front end error handling

+++ backend/routes/posts.js
@@ -34,9 +34,7 @@ const storage = multer.diskStorage({
 router.post(
   "",
   checkAuth,
-  multer({
-    storage: storage
-  }).single("image"),
+  multer({ storage: storage }).single("image"),
   (req, res, next) => {
     const url = req.protocol + "://" + req.get("host");
     const post = new Post({
@@ -45,24 +43,29 @@ router.post(
       imagePath: url + "/images/" + req.file.filename,
       creator: req.userData.userId
     });
-    post.save().then(createdPost => {
-      res.status(201).json({
-        message: "Post added successfully",
-        post: {
-          ...createdPost,
-          id: createdPost._id
-        }
+    post
+      .save()
+      .then(createdPost => {
+        res.status(201).json({
+          message: "Post added successfully",
+          post: {
+            ...createdPost,
+            id: createdPost._id
+          }
+        });
+      })
+      .catch(error => {
+        res.status(500).json({
+          message: "Creating a post failed!"
+        });
       });
-    });
   }
 );
 
 router.put(
   "/:id",
   checkAuth,
-  multer({
-    storage: storage
-  }).single("image"),
+  multer({ storage: storage }).single("image"),
   (req, res, next) => {
     let imagePath = req.body.imagePath;
     if (req.file) {
@@ -76,20 +79,19 @@ router.put(
       imagePath: imagePath,
       creator: req.userData.userId
     });
-    Post.updateOne({
-      _id: req.params.id,
-      creator: req.userData.userId
-    }, post).then(result => {
-      if (result.nModified > 0) {
-        res.status(200).json({
-          message: "Update successful!"
+    Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post)
+      .then(result => {
+        if (result.nModified > 0) {
+          res.status(200).json({ message: "Update successful!" });
+        } else {
+          res.status(401).json({ message: "Not authorized!" });
+        }
+      })
+      .catch(error => {
+        res.status(500).json({
+          message: "Couldn't udpate post!"
         });
-      } else {
-        res.status(401).json({
-          message: "Not authorized!"
-        });
-      }
-    });
+      });
   }
 );
 
@@ -112,36 +114,45 @@ router.get("", (req, res, next) => {
         posts: fetchedPosts,
         maxPosts: count
       });
+    })
+    .catch(error => {
+      res.status(500).json({
+        message: "Fetching posts failed!"
+      });
     });
 });
 
 router.get("/:id", (req, res, next) => {
-  Post.findById(req.params.id).then(post => {
-    if (post) {
-      res.status(200).json(post);
-    } else {
-      res.status(404).json({
-        message: "Post not found!"
+  Post.findById(req.params.id)
+    .then(post => {
+      if (post) {
+        res.status(200).json(post);
+      } else {
+        res.status(404).json({ message: "Post not found!" });
+      }
+    })
+    .catch(error => {
+      res.status(500).json({
+        message: "Fetching post failed!"
       });
-    }
-  });
+    });
 });
 
 router.delete("/:id", checkAuth, (req, res, next) => {
-  Post.deleteOne({
-    _id: req.params.id,
-    creator: req.userData.userId
-  }).then(result => {
-    if (result.n > 0) {
-      res.status(200).json({
-        message: "Deletion successful!"
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId })
+    .then(result => {
+      console.log(result);
+      if (result.n > 0) {
+        res.status(200).json({ message: "Deletion successful!" });
+      } else {
+        res.status(401).json({ message: "Not authorized!" });
+      }
+    })
+    .catch(error => {
+      res.status(500).json({
+        message: "Deleting posts failed!"
       });
-    } else {
-      res.status(401).json({
-        message: "Not authorized!"
-      });
-    }
-  });
+    });
 });
 
 module.exports = router;
